Commit Final?

+++ Microservices/Passenger/Controllers/PassengerController.cs
@@ -41,6 +41,7 @@ namespace PassengerAPI.Controllers
         public ActionResult<List<Passenger>> GetAllMinors()
         {
             var minors = _passengerService.GetAllMinors();
+            if (minors == null) return StatusCode(404, "Não existem passageiros menores de idade cadastrados!");
             return minors;
         }
 
@@ -51,7 +52,9 @@ namespace PassengerAPI.Controllers
         [HttpGet("Inactives")]
         public ActionResult<List<Passenger>> GetDeletedOnes()
         {
-            return Ok(_passengerService.GetDeletedOnes());
+            var inactives = _passengerService.GetDeletedOnes();
+            if (inactives == null) return StatusCode(404, "Não existem passageiros inativos!");
+            return inactives;
         }
 
 
@@ -61,7 +64,9 @@ namespace PassengerAPI.Controllers
         [HttpGet("RestrictedOnes")]
         public ActionResult<List<Passenger>> GetRestrictedOnes()
         {
-            return Ok(_passengerService.GetRestrictedOnes());
+            var restricted =_passengerService.GetRestrictedOnes();
+            if (restricted == null) return StatusCode(404, "Não existem passageiros restritos!");
+            return restricted;
         }
 
 
@@ -71,8 +76,8 @@ namespace PassengerAPI.Controllers
         [HttpGet("{cpf}")]
         public ActionResult<Passenger> GetByCPF(string cpf)
         {
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
             var passenger = _passengerService.GetByCPF(cpf);
-            if (!ValidateDocument.ValidateCPF(passenger.CPF, passenger.CPF)) return BadRequest("CPF Inválido!");
             if (passenger == null) return NotFound("Passageiro não encontrado!");
 
             return Ok(passenger);
@@ -86,7 +91,6 @@ namespace PassengerAPI.Controllers
         public ActionResult<Passenger> Post(PassengerDTO passengerDTO)
         {
             var dto = _postOffice.GetAddress(passengerDTO.CEP).Result;
-
             if (!ValidateDocument.ValidateCPF(passengerDTO.CPF, passengerDTO.CPF)) return BadRequest("CPF Inválido!");
 
             var number = 0;
@@ -121,11 +125,13 @@ namespace PassengerAPI.Controllers
 
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [HttpPut("/Address/{cpf}")]
-        public ActionResult<Passenger> UpdatePassengerAddress(string cpf, string cep)
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [HttpPut("/Address/number/complement/{cpf}")]
+        public ActionResult<Passenger> UpdatePassengerAddress(string cpf, string cep, int number, string complement)
         {
-            var passenger = _passengerService.UpdatePassengerAddress(cpf, cep);
-            if (passenger == null) return NotFound();
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
+            var passenger = _passengerService.UpdatePassengerAddress(cpf, cep, number, complement);
+            if (passenger == null) return NotFound("Passageiro não encontrado!");
 
             return Created(string.Empty, passenger);
         }
@@ -134,11 +140,13 @@ namespace PassengerAPI.Controllers
 
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpPut("/Street/{cpf}")]
         public ActionResult<Passenger> UpdatePassengerAddressStreet(string cpf, string streetName)
         {
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
             var passenger = _passengerService.UpdatePassengerAddressStreet(cpf, streetName);
-            if (passenger == null) return NotFound();
+            if (passenger == null) return NotFound("Passageiro não encontrado!");
 
             return Ok(passenger);
         }
@@ -147,11 +155,13 @@ namespace PassengerAPI.Controllers
 
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpPut("/Name/{cpf}")]
         public ActionResult<Passenger> UpdatePassengerName(string cpf, string name)
         {
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
             var passenger = _passengerService.UpdatePassengerName(cpf, name);
-            if (passenger == null) return NotFound();
+            if (passenger == null) return NotFound("Passageiro não encontrado!");
 
             return Ok(passenger);
         }
@@ -160,11 +170,13 @@ namespace PassengerAPI.Controllers
 
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpPut("/Gen/{cpf}")]
         public ActionResult<Passenger> UpdatePassengerGender(string cpf, char gen)
         {
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
             var passenger = _passengerService.UpdatePassengerGender(cpf, gen);
-            if (passenger == null) return NotFound();
+            if (passenger == null) return NotFound("Passageiro não encontrado!");
 
             return Ok(passenger);
         }
@@ -173,11 +185,13 @@ namespace PassengerAPI.Controllers
 
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpPut("/Phone/{cpf}")]
         public ActionResult<Passenger> UpdatePassengerPhone(string cpf, string phone)
         {
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
             var passenger = _passengerService.UpdatePassengerPhone(cpf, phone);
-            if (passenger == null) return NotFound();
+            if (passenger == null) return NotFound("Passageiro não encontrado!");
 
             return Ok(passenger);
         }
@@ -186,11 +200,13 @@ namespace PassengerAPI.Controllers
 
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpPut("/Status/{cpf}")]
         public ActionResult<Passenger> UpdatePassengerStatus(string cpf)
         {
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
             var passenger = _passengerService.UpdatePassengerStatus(cpf);
-            if (passenger == null) return NotFound();
+            if (passenger == null) return NotFound("Passageiro não encontrado!");
 
             return Ok(passenger);
         }
@@ -199,11 +215,13 @@ namespace PassengerAPI.Controllers
 
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpPut("/SetRestrict/{cpf}")]
         public ActionResult<Passenger> SetPassengerAsRestricted(string cpf)
         {
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
             var passenger = _passengerService.SetPassengerAsRestricted(cpf);
-            if (passenger == null) return NotFound();
+            if (passenger == null) return NotFound("Passageiro não encontrado!");
 
             return Ok(passenger);
         }
@@ -212,11 +230,13 @@ namespace PassengerAPI.Controllers
 
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpPut("/SetUnrestrict/{cpf}")]
         public ActionResult<Passenger> SetPassengerAsUnrestricted(string cpf)
         {
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
             var passenger = _passengerService.SetPassengerAsUnrestricted(cpf);
-            if (passenger == null) return NotFound();
+            if (passenger == null) return NotFound("Passageiro não encontrado!");
 
             return Ok(passenger);
         }
@@ -225,11 +245,13 @@ namespace PassengerAPI.Controllers
 
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpPut("/Reactivate/{cpf}")]
         public ActionResult<Passenger> ReactivatePassenger(string cpf)
         {
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
             var passenger =_passengerService.ReativatePassenger(cpf);
-            if (passenger == null) return NotFound();
+            if (passenger == null) return NotFound("Passageiro não encontrado!");
             return Ok(passenger);
         }
 
@@ -237,11 +259,13 @@ namespace PassengerAPI.Controllers
 
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpDelete("/Desactivate/{cpf}")]
         public ActionResult <Passenger>Delete(string cpf)
         {
+            if (!ValidateDocument.ValidateCPF(cpf, cpf)) return BadRequest("CPF Inválido!");
             var passenger = _passengerService.Delete(cpf);
-            if (passenger == null) return NotFound();
+            if (passenger == null) return NotFound("Passageiro não encontrado!");
             return Ok(passenger);
         }
     }
